gameobjects

+++ Game/Scripts/Collision.cs
@@ -8,7 +8,7 @@ namespace Game
 {
     public static class Collision
     {
-        public static bool RectRect(int rect1X, int rect1Y, int rect1Size, int rect2X, int rect2Y, int rect2Size)
+        public static bool RectRect(float rect1X, float rect1Y, float rect1Size, float rect2X, float rect2Y, float rect2Size)
         {
             return (rect1X < rect2X + rect2Size &&
                     rect1Y < rect2Y + rect2Size &&
